feat(search): home hero on shared SearchBar; mobile overlay stacked; retire SearchForm
Slices 4, 6, 7 + two fixes found during verification.

- HomeSearchBar: home container with the FilterModal/useBatchedFilters
  staging machinery lifted from SearchForm around the shared bar
  (externalBudget keeps the pill's Min/Max and the drawer's price slider on
  one source of truth). HomeClient lazy-loads it; Suspense fallback matches
  the new 68px geometry.
- MobileSearchOverlay: form block replaced by the stacked shared bar
  (mobile- id prefix; real form so Enter submits; gains env-gated What field
  and min>max swap). Recents list and portal/FocusTrap shell unchanged.
- Deleted SearchForm.tsx (+ its test; HomeSearchBar.test.tsx ports the
  behavioral contract, 82 tests) and CompactSearchPill.tsx (no importers);
  stale comments updated.
- Fix: duplicate-submit dedupe ran after cancelling the pending debounced
  navigation, swallowing the search entirely (latent SearchForm bug);
  dedupe now checked first, with a regression test.
- Fix: SearchBarScrim portal rendered during hydration caused React #418 on
  /search (server null vs client portal); now mount-gated. Verified gone on
  a prod build.
