+++ src/components/search/SearchBar/SearchBarScrim.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useEffect, useState } from "react";
 import { createPortal } from "react-dom";
 import { cn } from "@/lib/utils";
 
@@ -13,9 +14,19 @@ interface SearchBarScrimProps {
  * below the fixed header (z-1100) but above the page content. Purely visual
  * chrome — the document-level outside-click/Escape listeners remain the
  * actual collapse mechanism; clicking the scrim is just one such outside click.
+ *
+ * Rendered only after mount: the server has no portal target, and returning
+ * null on the server but a portal during hydration is a React 19 hydration
+ * mismatch (#418). The scrim is invisible at rest, so mount-gating is free.
  */
 export function SearchBarScrim({ visible, onDismiss }: SearchBarScrimProps) {
-  if (typeof document === "undefined") return null;
+  const [mounted, setMounted] = useState(false);
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  if (!mounted) return null;
 
   return createPortal(
     <div
