+++ src/components/search/index.ts
@@ -2,7 +2,7 @@
  * Search UI Components
  *
  * Presentational components for the search page filter UI.
- * All filter state and URL-sync logic remains in SearchForm.tsx.
+ * All filter state and URL-sync logic remains in SearchBar/HomeSearchBar.tsx.
  */
 
 export { FilterPill } from "./FilterPill";
